lib: sbi: Probe pmp_count in read-only
Some CPU implementations such as T-Head C908 will raise illegal
instruction exception when writing a different value to some pmpaddr
which is locked.

To avoid this, a solution is to probe pmp_count in read-only. This will
not affect the existing CPU implementation as RISC-V Privileged
Architecture specified that access to a non-existent CSR raises an
illegal instruction exception.

Signed-off-by: Yangyu Chen <cyy@cyyself.name>