lib: sbi: Probe reserved PMP regions before OpenSBI start
Some platform which has different domains for different CPU cores may
disable some device by PMP in the bootloader.

For example, Canaan Kendryte K230 provides SDK which locks some PMP
entries in M-Mode U-Boot[1]. If the OpenSBI did not probe it, OpenSBI
may get the wrong result from hart_pmp_get_allowed_addr() as the
pmpaddr0 can not be changed if it is locked. It will also break the
sbi_hart_pmp_configure() as it assumes the useable PMP index starts
from 0.

Thus, providing a function to probe the PMP regions that are reserved
before OpenSBI starts and store the information in sbi_hart_features
will be helpful to avoid the issues mentioned above.

[1] https://github.com/kendryte/k230_sdk/blob/72be167ad4b3a7e91f7cc6abbe576d347b885af4/src/little/uboot/arch/riscv/cpu/k230/cpu.c#L82

Co-developed-by: Vivian Wang <dramforever@live.com>
Co-developed-by: Wende Tan <twd2.me@gmail.com>
Signed-off-by: Vivian Wang <dramforever@live.com>
Signed-off-by: Wende Tan <twd2.me@gmail.com>
Signed-off-by: Yangyu Chen <cyy@cyyself.name>